Add: FUNDING.yml
This commit introduces a FUNDING.yml file to the repository. The FUNDING.yml file is used to specify a way for users to financially support the development and maintenance of this project. By adding this file, we aim to provide a clear and accessible way for contributors and users to show their support.

The following funding option has been added:
- GitHub Sponsors: https://github.com/sponsors/gayanukabulegoda

This funding option will be displayed in the repository's sidebar, making it easy for interested parties to contribute. We appreciate any and all support from our community, as it helps us continue to improve and maintain this project.

Thank you for your support!